Remove project module
